+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- #10: Add links to issue numbers in Pull Request titles on the PR page
+
 ### Fixed
 - #13: Bugzilla bugs whose info could not be fetched will appear with the
 correct bug number